api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.0
openai==1.59.5
qdrant-client==1.12.1
sentence-transformers==3.3.1
psycopg[binary]==3.3.2
pydantic==2.10.5
clerk-backend-api==1.5.0
pyjwt==2.9.0
cryptography>=43.0.1,<44.0.0
